Add utils folder with form values + move form to new component
