+++ frontend/src/metabase/containers/SaveQuestionModal/index.ts
@@ -1,1 +1,2 @@
 export * from "./SaveQuestionModal";
+export { SAVE_QUESTION_SCHEMA } from "metabase/containers/SaveQuestionModal/schema";
